added undo logic

+++ src/render.js
@@ -59,18 +59,83 @@ function deleteShipSelection() {
   }
 }
 // clear the background for every block in the grid
-function clearGrid(removeShips = false) {
+function clearGrid(removeShips = false, removeBlockZones = false) {
   const selector = displayController.board;
   const children = selector.childNodes;
   children.forEach((childElem) => {
     childElem.style.background = 'none';
-    if (removeShips) {
+    if (removeBlockZones || removeShips) {
       childElem.classList.remove('blockedzone');
       childElem.classList.add('dropzone');
-      childElem.innerHTML = '';
-      deleteShipSelection(); // rerender ship select
-      createShipSelection([0, 4, 3, 2, 1]);
+      if (removeShips) {
+        childElem.innerHTML = '';
+        deleteShipSelection(); // rerender ship select
+        createShipSelection([0, 4, 3, 2, 1]);
+      }
     }
+  });
+}
+
+function removeShip(parent, child) {
+  const activeGrids = JSON.parse(child.dataset.activegrids);
+  activeGrids.forEach((id) => {
+    const selector = document.getElementById(`grid${id}`);
+    selector.classList.remove('shipzone');
+    selector.classList.add('dropzone');
+  });
+  parent.removeChild(child);
+}
+
+function grabGridID(e) {
+  let gridID = e.id.replace(/[^0-9]/g, '');
+  gridID = parseInt(gridID, 10);
+  return (gridID);
+}
+function getShipZones() {
+  const shipZoneArray = [];
+  displayController.board.childNodes.forEach((grid) => {
+    if (grid.classList.contains('shipzone')) {
+      shipZoneArray.push(grid);
+    }
+  });
+  return (shipZoneArray);
+}
+function getBlockedZones(elemArray) {
+  const blockedArray = []; // collect elements for the blocked grids
+  let borderLeft = false; // bool to check if a ship grid is on the left or right border
+  let borderRight = false;
+  // loop through the grid spaces where the ship is
+  elemArray.forEach((elem) => {
+    const elemID = grabGridID(elem);
+    if (((elemID % 10) - 1) === 0) { borderLeft = true; } // check if ship is placed on the border
+    if (elemID % 10 === 0) { borderRight = true; }
+    const offsetModifiers = [-10, 0, 10]; // modifiers go up and down the grids
+    // for each modifier, use a for loop to append blocked grids to the blocked array
+    offsetModifiers.forEach((modifier) => {
+      // i will range from -1 to 1 and will correspond to the left middle and right
+      for (let i = -1; i < 2; i++) {
+        // makesure that grids outside of the border arent added
+        if (!(((i === 1) && (borderRight)) || ((i === -1) && (borderLeft)))) {
+          let blockedID = elemID + modifier;
+          blockedID += i;
+          const blockedGrid = document.getElementById(`grid${String(blockedID)}`);
+          // checks to make sure the blockedGrid isnt already in the array, or the ship grids
+          // make sure that no grids are added outside of the container (>1 or <100)
+          if (!blockedArray.includes(blockedGrid) && !elemArray.includes(blockedGrid)
+           && blockedID <= 100 && blockedID >= 1) {
+            blockedArray.push(blockedGrid);
+          }
+        }
+      }
+    });
+  });
+  return blockedArray;
+}
+
+function blockZones(blockedArray) {
+  blockedArray.forEach((blockedGrid) => {
+    blockedGrid.classList.remove('dropzone');
+    blockedGrid.classList.add('blockedzone');
   });
 }
 
@@ -98,15 +163,18 @@ function undo(btnSelector) {
 
       countArray[shipSize] += 1;
       console.log(blockedZones);
-      blockedZones.forEach((gridID) => {
-        const selector = document.getElementById(`grid${gridID}`);
-        selector.classList.remove('blockedzone');
-        selector.classList.add('dropzone');
-      });
+      // blockedZones.forEach((gridID) => {
+      //   const selector = document.getElementById(`grid${gridID}`);
+      //   selector.classList.remove('blockedzone');
+      //   selector.classList.add('dropzone');
+      // });
 
       deleteShipSelection();
       createShipSelection(countArray);
-      gridSelector.removeChild(shipsSelector);
+      removeShip(gridSelector, shipsSelector);
+      clearGrid(false, true);
+      const shipZones = getShipZones();
+      blockZones(getBlockedZones(shipZones));
 
       displayController.board.dataset.history = JSON.stringify(historyArray);
       displayController.shipSelectContainer.dataset.counter = JSON.stringify(countArray);
@@ -154,5 +222,6 @@ function createBoardBtns() {
 }
 
 export {
-  createGridBlocks, createShipSelection, deleteShipSelection, createBoardBtns, clearGrid,
+  createGridBlocks, createShipSelection, deleteShipSelection,
+  createBoardBtns, clearGrid, blockZones, getBlockedZones,
 };
